Register the event only on Paper servers of version 1.21.9 and above
