+++ src/main/java/com/bekvon/bukkit/residence/listeners/ResidenceListener1_21_9_Paper.java
@@ -1,0 +1,45 @@
+package com.bekvon.bukkit.residence.listeners;
+
+import org.bukkit.entity.Entity;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+
+import com.bekvon.bukkit.residence.Residence;
+import com.bekvon.bukkit.residence.containers.Flags;
+import com.bekvon.bukkit.residence.protection.FlagPermissions;
+
+import net.Zrips.CMILib.Entities.CMIEntityType;
+
+import io.papermc.paper.event.entity.ItemTransportingEntityValidateTargetEvent;
+
+public class ResidenceListener1_21_9_Paper implements Listener {
+
+    private Residence plugin;
+
+    public ResidenceListener1_21_9_Paper(Residence plugin) {
+        this.plugin = plugin;
+    }
+
+    @EventHandler(priority = EventPriority.LOWEST)
+    public void onCopperGolemInteract(ItemTransportingEntityValidateTargetEvent event) {
+
+        Entity entity = event.getEntity();
+        if (entity == null)
+            return;
+
+        // disabling event on world
+        if (plugin.isDisabledWorldListener(entity.getWorld()))
+            return;
+
+        if (CMIEntityType.get(entity) != CMIEntityType.COPPER_GOLEM)
+            return;
+
+        FlagPermissions perms = FlagPermissions.getPerms(event.getBlock().getLocation());
+        if (perms.has(Flags.golemopenchest, perms.has(Flags.container, true)))
+            return;
+
+        event.setAllowed(false);
+
+    }
+}
